confirmacion de alerta

+++ frontend/src/app/modules/home/components/detalles/detalles.component.ts
@@ -8,6 +8,9 @@ import { ApiService } from 'src/app/shared/services/api/api.service';
 import { Router } from '@angular/router';
 import { delay } from 'rxjs/operators';
 
+import Swal from 'sweetalert2';
+
+
 @Component({
   selector: 'app-detalles',
   templateUrl: './detalles.component.html',
@@ -38,13 +41,32 @@ export class DetallesComponent implements OnInit {
   }
 
   deleteTrailler(){
-    this.api.deleteTraillers(this.detalle.trailer.trailler.id).subscribe(resp=>{
-      console.log(resp);
-    });   
-    setTimeout(() => 
-    {
-      location.reload();
-    },400);
+    Swal.fire({
+      title: 'Estas seguro?',
+      text: "Una vez eliminado no se podra recuperar!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Borrar Trailer'
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.api.deleteTraillers(this.detalle.trailer.trailler.id).subscribe(resp=>{
+          console.log(resp);
+        });   
+        setTimeout(() => 
+        {
+          location.reload();
+        },400);
+        Swal.fire(
+          'Deleted!',
+          'El trailer ha sido eliminado.',
+          'success'
+        )
+      }else{
+        Swal.fire("No se elimino el trailer!");
+      }
+    });
     
   }
